Updated to use MQ as the transport.
This has been tested via the `mosquitto_pub` command; inject
a full-request such as:

   GET / HTTP/1.0
   Host: foo.example.com

The result is that the response is sent back down the same
channel, as expected.

Porting the server should be pretty simple, based on this work.

Since the HTTP-server and the MQ-server are on the same host
I'm going to cheat and only talk to MQ on the localhost for
the server.